ai-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
ultralytics>=8.3.0
mediapipe==0.10.33
skin-tone-classifier==1.2.6
google-generativeai==0.8.0
python-multipart==0.0.12
httpx==0.27.0
python-dotenv==1.0.1
Pillow>=11.0.0
numpy>=1.23.0
opencv-python-headless>=4.10.0
scikit-learn==1.6.1
anthropic>=0.78.0
